Players join tables of their own accord now - not bet based - and the remote updates are done entirely through the Attendant :)

+++ src/main/scala/code/lib/bj/card/Shoe.scala
@@ -18,7 +18,7 @@ package lib
 package bj.card
 
 /** This class implements the shoe from which we deal cards. */
-class Shoe(seed : Int) {
+class Shoe(seed : Long) {
     /** Number of decks in the shoe */
     val DECK_COUNT = 6
